build: remove irrelevant nuttx version generation
Part of the build, NuttX issues a git command to get the current version,
includes it in .version, and converts that into a header (nuttx/version.h). This
whole version computation and the resulting header are not used in the firmware.

Let's remove it since we already have our own version computation.

Testing Done: compiled
Bug Number: SW-2021
Review URL:
Reviewed by:

Signed-off-by: Joel Porquet <joel@porquet.org>